test(runtime): isolate FOR EACH control-flow regressions
