+++ tests/CMakeLists.txt
@@ -5523,6 +5523,7 @@ copperfin_add_test(test_prg_engine_control_flow cf_xbase_runtime
     test_prg_engine_control_flow.cpp
     test_prg_engine_control_flow_shared_test_helpers.cpp
     test_prg_engine_control_flow_control_flow_basics.cpp
+    test_prg_engine_control_flow_for_each.cpp
     test_prg_engine_control_flow_array_copy.cpp
     test_prg_engine_control_flow_implicit_scope.cpp
     test_prg_engine_control_flow_variables_and_scope.cpp
